// frontend/src/utilities/types/CalculateTotalProductPriceTypes.ts
/* eslint-disable typescript-sort-keys/interface */
/* eslint-disable id-length */

export type PriceVariablesData = {
  backNumberOfColors: number;
  frontNumberOfColors: number;
  productId: number;
  quantityBySizes: {
    XS: 0;
    S: 0;
    M: 0;
    LG: 0;
    XL: 0;
    '2XL': 0;
    '3XL': 0;
    '4XL': 0;
    '5XL': 0;
  };
};